ART: fix overflow in weighted allocated bytes metrics
Test: Run art with -XX:DumpGCPerformanceOnShutdown on some benchmarks.
Bug: 112187497

+++ runtime/gc/heap.cc
@@ -214,7 +214,7 @@ Heap::Heap(size_t initial_size,
       long_gc_log_threshold_(long_gc_log_threshold),
       process_cpu_start_time_ns_(ProcessCpuNanoTime()),
       last_process_cpu_time_ns_(process_cpu_start_time_ns_),
-      weighted_allocated_bytes_(0u),
+      weighted_allocated_bytes_(0.0),
       ignore_max_footprint_(ignore_max_footprint),
       zygote_creation_lock_("zygote creation lock", kZygoteCreationLock),
       zygote_space_(nullptr),
@@ -1072,7 +1072,7 @@ void Heap::RemoveSpace(space::Space* space) {
 void Heap::CalculateWeightedAllocatedBytes() {
   uint64_t current_process_cpu_time = ProcessCpuNanoTime();
   uint64_t bytes_allocated = GetBytesAllocated();
-  uint64_t weight = current_process_cpu_time - last_process_cpu_time_ns_;
+  double weight = current_process_cpu_time - last_process_cpu_time_ns_;
   weighted_allocated_bytes_ += weight * bytes_allocated;
   last_process_cpu_time_ns_ = current_process_cpu_time;
 }
